chore: move construct_leaf_node to leaf_node.rs
